requirements.txt
streamlit
python-dotenv
google-generativeai
wordcloud
textblob
matplotlib